Relayout all spec files
