+++ cypress/integration/application(Not Pages)/index.js
