+++ cypress/integration/pages/articlePages/needsRefactoringIntoParentDir/newsArticleAmpSpec.js
@@ -1,20 +1,15 @@
-import services from '../support/config/services';
-import { describeForLocalOnly } from '../support/limitEnvRuns';
+import services from '../../../../support/config/services';
 
-// TODO Enable all disabled tests below once bbc/simorgh#1906 has been merged.
-//   The Metadata container performs some AMP work, so the AMP on the front page
-//   will not be valid until the Metadata container has been added into the front
-//   page container. 🙃
-describeForLocalOnly('AMP Tests on a .amp page', () => {
+describe('AMP Tests on a .amp page', () => {
   // eslint-disable-next-line no-undef
   before(() => {
-    cy.visit(`${services.igbo.pageTypes.frontPage}.amp`);
+    cy.visit(`/news/articles/${services.news.pageTypes.articles.asset}.amp`);
   });
 
   describe('AMP Status', () => {
     it('should return a 200 response', () => {
       cy.testResponseCodeAndType(
-        `${services.igbo.pageTypes.frontPage}.amp`,
+        `/news/articles/${services.news.pageTypes.articles.asset}.amp`,
         200,
         'text/html',
       );
@@ -23,30 +18,34 @@ describeForLocalOnly('AMP Tests on a .amp page', () => {
 
   it('should error gracefully', () => {
     cy.testResponseCodeAndType(
-      `${services.igbo.pageTypes.frontPage}.cake`,
+      `/news/articles/${services.news.pageTypes.articles.asset}.cake`,
       404,
       'text/html',
     );
     cy.testResponseCodeAndType(
-      `/amp${services.igbo.pageTypes.frontPage}`,
+      `/news/lol/${services.news.pageTypes.articles.asset}.amp`,
       404,
       'text/html',
     );
     cy.testResponseCodeAndType(
-      `${services.igbo.pageTypes.frontPage}/amp`,
+      `/cake/articles/${services.news.pageTypes.articles.asset}.amp`,
       404,
       'text/html',
     );
-    cy.testResponseCodeAndType(`/cake.amp`, 404, 'text/html');
   });
 
-  xit('should have AMP attribute', () => {
+  it('should have AMP attribute', () => {
     cy.get('html').should('have.attr', 'amp');
   });
 
+  it('should have lang and dir attributes', () => {
+    cy.hasHtmlLangDirAttributes({ lang: 'en-gb', dir: 'ltr' });
+  });
+
   it('should load the AMP framework', () => {
-    // .eq(1) gets the amp <script> as:
+    // .eq(2) gets the amp <script> as:
     // the first loaded is a Cypress <script>
+    // the second loaded is the Schema.org metadata script
     cy.get('head script')
       .eq(2)
       .should('have.attr', 'src', 'https://cdn.ampproject.org/v0.js');
@@ -66,6 +65,14 @@ describeForLocalOnly('AMP Tests on a .amp page', () => {
         'src',
         'https://cdn.ampproject.org/v0/amp-consent-0.1.js',
       );
+
+    cy.get('head script')
+      .eq(5)
+      .should(
+        'have.attr',
+        'src',
+        'https://cdn.ampproject.org/v0/amp-analytics-0.1.js',
+      );
   });
 
   it('should load the AMP body scripts', () => {
@@ -83,30 +90,27 @@ describeForLocalOnly('AMP Tests on a .amp page', () => {
       .should('be', 2); // 1 for amp-geo + 1 for amp-consent
     cy.get('head script')
       .its('length')
-      .should('be', 4); // 1 for amp.js + 1 for amp-geo + 1 for amp-consent + 1 that Cypress injects into the head
+      .should('be', 5); // 1 for amp.js + 1 for amp-geo + 1 for amp-consent + 1 for amp-analytics + 1 that Cypress injects into the head
   });
 
   it('should contain an amp-img', () => {
-    cy.get('li')
+    cy.get('figure')
+      .eq(0)
       .should('be.visible')
       .within(() => {
         cy.get('amp-img').should('be.visible');
       });
   });
 
-  xit('should include the canonical URL', () => {
-    const { origin } = window.location;
-    const canonicalOrigin = origin.includes('localhost')
-      ? 'https://www.bbc.com'
-      : origin;
-
+  it('should include the canonical URL', () => {
+    const canonicalOrigin = 'https://www.bbc.com';
     cy.checkCanonicalURL(
-      `${canonicalOrigin}${services.igbo.pageTypes.frontPage}`,
+      `${canonicalOrigin}/news/articles/${services.news.pageTypes.articles.asset}`,
     );
   });
 
-  xit('should not have an AMP attribute on the main article', () => {
-    cy.visit(`${services.igbo.pageTypes.frontPage}`);
+  it('should not have an AMP attribute on the main article', () => {
+    cy.visit(`/news/articles/${services.news.pageTypes.articles.asset}`);
     cy.get('html').should('not.have.attr', 'amp');
   });
 });
