+++ cypress/integration/shouldBeRefactoredIntoOtherFolders/scriptSpec.js
@@ -1,4 +1,4 @@
-import config from '../support/config/services';
+import config from '../../support/config/services';
 
 const host = 'http://localhost.bbc.com:7080';
